Bug #34 - fix attempted completion method return value
When a custom completion_proc is set, Readline would *look* like the
completion_append_character was not being appended to the match. For
example, from the bug[1]:

  Readline.completion_proc = Proc.new { |word| [ word ] }

  abc<TAB>  becomes:  abc      <-- rather than:  abc<SPACE>
  <TAB>     becomes:  nothing  <-- rather than:  <SPACE>

In fact, Readline.readline_attempted_completion_function was returning
an incorrect array.

This patch adds tests to cover this bug and exercise the case folding
feature of that method.

1 - https://github.com/luislavena/rb-readline/issues/34

Tested-by: Mark Somerville <mark@scottishclimbs.com> [MRI 1.9.2, MRI 1.8.7]
Signed-off-by: Mark Somerville <mark@scottishclimbs.com>